consumer/requirements.txt
aio-pika==9.4.1
redis==5.0.8
python-dotenv==1.0.1
